bpf: Allow from host with remote-node-identity=false
According to the documentation:

    One can set enable-remote-node-identity=false in the ConfigMap to
    retain the Cilium 1.6.x behavior.

The above is true when evaluating policy, but not true from the
DROP_INVALID_IDENTITY perspective as during an upgrade from v1.6 to
v1.7, v1.6 nodes may send traffic to v1.7 nodes with the "host" identity
and the v1.7 nodes will drop such traffic with this error code. Mitigate
this by also covering this datapath case with the
`EnableRemoteNodeIdentity` flag check.

Signed-off-by: Joe Stringer <joe@cilium.io>