daemon: Add hidden allow-remote-src flag
This flag allows to override the default setting for whether to accept
traffic into a node if the source node reports the source identity as
"host". Default is "auto" - to allow if enableRemoteNodeIdentity is
disabled; deny if enableRemoteNodeIdentity is enabled. If specified
directly, it will take on the configured behaviour (allow such traffic
if true, otherwise drop it).

Signed-off-by: Joe Stringer <joe@cilium.io>